Update bundle_ru.properties
Hi, it's me, Szaman PL / Sir Wonder from official Mirage discord server. In this proposal I checked and corrected what was necessary to correct. Text after "##" is just comment to seperate stuff as new text to translate is available, so it don't need translating as it don't appear in the game.